+++ src/main/kotlin/com/sendy/infrastructure/persistence/UserEntity.kt
@@ -1,6 +1,7 @@
 package com.sendy.infrastructure.persistence
 
 import com.sendy.support.util.Aes256Converter
+import com.sendy.user.application.dto.UpdateUserRequestDto
 import jakarta.persistence.*
 import java.time.LocalDateTime
 
@@ -11,6 +12,8 @@ import java.time.LocalDateTime
 @Entity
 @Table(name = "users")
 class UserEntity(
+
+
     id: Long,
 
     @Column(name = "name", length = 50, nullable = false)
@@ -48,4 +51,37 @@ class UserEntity(
 
     @Column(name = "email_verified", nullable = false, columnDefinition = "BOOLEAN DEFAULT false")
     var emailVerified: Boolean = false
-) : Identity(id)
+) : Identity(id){
+        fun update(updateDto : UpdateUserRequestDto): UserEntity {
+        return UserEntity(
+            id = this.id,
+            password = this.password,
+            name = if (updateDto.name == null) this.name else updateDto.name,
+            phoneNumber = if(updateDto.phoneNumber == null) this.phoneNumber else updateDto.phoneNumber,
+            email = if(updateDto.email == this.email) this.email else updateDto.email,
+            ci = this.ci,
+            birth = if(updateDto.birth == null) this.birth else updateDto.birth,
+            isDelete = this.isDelete,
+            emailVerified = this.emailVerified,
+            createAt = this.createAt,
+            updateAt = LocalDateTime.now(),
+            deleteAt = this.deleteAt,
+        )
+    }
+        fun deleteUser(): UserEntity {
+        return UserEntity(
+            id = this.id,
+            password = this.password,
+            name = this.name,
+            phoneNumber = this.phoneNumber,
+            email = this.email,
+            ci = this.ci,
+            birth = this.birth,
+            isDelete = true,
+            emailVerified = this.emailVerified,
+            createAt = this.createAt,
+            updateAt = this.createAt,
+            deleteAt = LocalDateTime.now(),
+        )
+    }
+}
