+++ src/main/resources/application.yml
@@ -55,7 +55,7 @@ springdoc:
     display-request-duration: true
     operations-sorter: alpha
 
-jwt:
+token:
   secret-key: DYS6feNXKYnsLOebpSU0Ljm05bgit9j3sc3YYoBYh2E=
   access-token-expire-time: 300
   refresh-token-expire-time: 7200
